+++ src/test/java/org/morphix/reflection/MethodsGetFunctionalInterfaceMethodTest.java
@@ -51,6 +51,7 @@ class MethodsGetFunctionalInterfaceMethodTest {
 
 	public static interface C {
 		void foo();
+
 		void goo();
 	}