+++ src/test/java/org/morphix/reflection/ReflectionGetClassWithPrefixTest.java
@@ -42,6 +42,7 @@ class ReflectionGetClassWithPrefixTest {
 	@Test
 	void shouldThrowExceptionIfClassWithPrefixDoesNotExist() {
 		ReflectionException e = assertThrows(ReflectionException.class, () -> Reflection.getClassWithPrefix(Int.class, "P"));
-		assertThat(e.getMessage(), equalTo("Could not find class with prefix '" + ReflectionGetClassWithPrefixTest.class.getCanonicalName() + "$PInt'"));
+		assertThat(e.getMessage(),
+				equalTo("Could not find class with prefix '" + ReflectionGetClassWithPrefixTest.class.getCanonicalName() + "$PInt'"));
 	}
 }