+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ---
 
+`1.0.8`
+
+Added `Reflection.isClassPresent` to test if a class is present in the class path.
+
+---
+
 `1.0.7`
 
 - Renamed all `Fields` methods to follow the pattern `getAll*` for methods returning lists and `getOne*` for methods returning single fields.